More elegant and more functional way to make 2.2 Ghz frequence available, ~2000 score increased in Antutu Benchmark.
In this way the cpu is detected as CPU_LEVEL_2 and this make available all scripts related to it as conceived from mediatek. In other words, to make 2.2 ghz available but left the cpu to be detected as cpu_level_1 does not make scripts related to pll (frequence handling) and temperature handling working correctly as programmed.